move code systemuser frontend code to amUI folder
