+++ src/features/amUI/systemUser/components/DelegationCheckError/DelegationCheckError.module.css
