+++ src/features/amUI/systemUser/components/RightsList/RightsList.module.css
