+++ src/features/amUI/systemUser/components/RightsList/RightsList.tsx
